+++ packages/language-core/src/utils/formatUsingComponentsPath.ts
@@ -1,4 +1,5 @@
 import type * as ts from 'typescript'
+import * as fs from 'node:fs'
 import * as path from 'node:path'
 import { URI } from 'vscode-uri'
 import { findResult } from './utils'
@@ -7,11 +8,14 @@ import { tryResolveByTsConfig, tryResolvePackage } from './resolve'
 export const uriToFileName = (uri: string) =>
   URI.parse(uri).fsPath.replace(/\\/g, '/')
 
+// TODO
+// 目前仅支持自动寻找补全 .mpx 后缀路径
+// .ios.mpx 和 .ali.mpx 这种条件编译后缀，暂时不支持，因为需要根据编译时环境变量来判断
 export async function formatUsingComponentsPath(
   componentPath: string = '',
   uri: string,
   compilerOption: ts.CompilerOptions,
-): Promise<string | undefined> {
+) {
   if (!componentPath) return
 
   const queryIndex = componentPath.indexOf('?')
@@ -21,13 +25,35 @@ export async function formatUsingComponentsPath(
 
   let formattedFilePath = ''
 
-  if (componentPath.startsWith('./') || componentPath.startsWith('../')) {
-    // 目前仅支持自动寻找补全 .mpx 后缀路径
-    // .ios.mpx 和 .ali.mpx 这种条件编译后缀，暂时不支持，因为需要根据编译时环境变量来判断
-    if (!componentPath.endsWith('.mpx')) {
-      componentPath += '.mpx'
+  const isRealFile = (filePath: string) => {
+    try {
+      const stats = fs.statSync(filePath, { throwIfNoEntry: false })
+      return stats && stats.isFile()
+    } catch {
+      return false
     }
-    formattedFilePath = path.join(uriToFileName(uri), '..', componentPath)
+  }
+
+  if (componentPath.startsWith('./') || componentPath.startsWith('../')) {
+    const basePath = path.join(uriToFileName(uri), '..', componentPath)
+
+    if (!componentPath.endsWith('.mpx')) {
+      const filePath = basePath + '.mpx'
+      if (isRealFile(filePath)) {
+        formattedFilePath = filePath
+      } else {
+        return { error: true }
+      }
+    } else {
+      // double check, eg: './list.mpx' -> './list.mpx.mpx'
+      if (isRealFile(basePath)) {
+        formattedFilePath = basePath
+      } else if (isRealFile(basePath + '.mpx')) {
+        formattedFilePath = basePath + '.mpx'
+      } else {
+        return { error: true }
+      }
+    }
   } else {
     const result = await findResult(
       [
@@ -37,9 +63,11 @@ export async function formatUsingComponentsPath(
       fn => fn(),
     )
     if (result) {
-      return result
+      formattedFilePath = result
+    } else {
+      return { error: true }
     }
   }
 
-  return formattedFilePath
+  return { result: formattedFilePath }
 }