+++ packages/language-service/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/mpx-ecology/language-tools/compare/v1.0.4...v1.2.0) (2025-07-16)
+
+
+### Features
+
+* enhance json usingComponents paths handler and add errors diagnostics for mpx json ([a0bc4dc](https://github.com/mpx-ecology/language-tools/commit/a0bc4dcd6f05b32d533124229c16f59f03c37783))
+* **service:** support schema for mpx json ([06fcfa2](https://github.com/mpx-ecology/language-tools/commit/06fcfa29348d389d1ce7cffbc5a9c6739bde6e1d))
+* **service:** support usingComponents document link for json-js ([a819598](https://github.com/mpx-ecology/language-tools/commit/a8195986d14b8c43cfd3599b17e99fd7cc2c3fcc))
+* **service:** update json schema ([4479a50](https://github.com/mpx-ecology/language-tools/commit/4479a50316969b6a621deab07d49119ae419c536))
+
+
+
+
+
+## [1.0.4](https://github.com/mpx-ecology/language-tools/compare/v1.0.3...v1.0.4) (2025-07-15)
+
+
+### Features
+
+* **service:** update attrs info ([701fea6](https://github.com/mpx-ecology/language-tools/commit/701fea611252df6891f00336b6fe85ed31fab34d))
+* **service:** wip add the attributes of the tag ([55a9266](https://github.com/mpx-ecology/language-tools/commit/55a9266618dc1983b98f671d8337cf29af00f1aa))
+* **service:** wip add the attributes of the tag ([758f164](https://github.com/mpx-ecology/language-tools/commit/758f164f60544db4aef16994bd6d0802aa8ba509))
+
+
+
+
+
 ## [1.0.3](https://github.com/mpx-ecology/language-tools/compare/v1.0.2...v1.0.3) (2025-07-14)
 
 **Note:** Version bump only for package @mpxjs/language-service